// src/types/diamonds/Enums.ts
export enum Cut {
  Good = 1,
  VeryGood = 2,
  Excellent = 3,
}
export enum Clarity {
  S12 = 1,
  S11 = 2,
  VS2 = 3,
  VS1 = 4,
  VVS2 = 5,
  VVS1 = 6,
  IF = 7,
  FL = 8,
}
export enum Color {
  K = 1,
  J = 2,
  I = 3,
  H = 4,
  G = 5,
  F = 6,
  E = 7,
  D = 8,
}
export enum Polish {
  Poor = 1,
  Fair = 2,
  Good = 3,
  VeryGood = 4,
  Excellent = 5,
}
export enum Symmetry {
  Poor = 1,
  Fair = 2,
  Good = 3,
  VeryGood = 4,
  Excellent = 5,
}
export enum Girdle {
  ExtremelyThin = 1,
  VeryThin = 2,
  Thin = 3,
  Medium = 4,
  SlightlyThick = 5,
  Thick = 6,
  VeryThick = 7,
  ExtremelyThick = 8,
}
export enum Culet {
  None = 1,
  Very_Small,
  Small,
  Medium,
  Slightly_Large,
  Large,
  Very_Large,
  Extremely_Large,
}
export enum Fluorescence {
  None = 1,
  Faint,
  Medium,
  Strong,
}